requirements.txt
fastapi==0.115.3
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
PyJWT==2.9.0
SQLAlchemy==2.0.34
pytest==7.4.4
httpx==0.27.0
redis>=5.0
rq>=1.16
cryptography>=42.0